updating SwiftPM to distribute a binary target
Additionally we are updating build_xcframework.sh to
automatically archive the XCFramework and update the Swift.package
Checksum to the checksum corresponding to the new archive. It also
asks for the new version number and updates the URL in the Swift
package to correspond to the new release.
Note that this uses a 'dumb' SED command that replaces line 16 and 17 of
Swift.package so changes to that file could break this command.