checking for end_time before displaying IAM
This change involves reading in end_time from JSON when creating an IAM
so that we can decide to not show the IAM (Cached case) if the IAM
should be finished. The date comes in iso8601 format so I added a
NSDateFormatter category with that format to easily retrieve it.
Note we cannot use the Apple provided iso8601 date formatter class until
we drop iOS 9 support.