Removing NSUserDefaultsOverrider
Now that our unit tests are run inside a test app we no longer need to mock NSUserDefaults. This gets us detection of bad storage like trying to store a custom object in a dictionary instead of as a codeable.

This commit also fixes the crash in InAppMessagingIntegrationTests by changing save/getDictionary to save/getCodeableData