Implementing didReceiveNotificationRequest with contentHandler
We will call the contentHandler with the replacement content to display the notification, but we will not return from the function until confirmed deliveries are completed. This allows us to delay the confirmed delivery server call since the NSE will not be killed until the function finishes.

Using dispatch_semaphore in the success/failure block to signal completion.