Unit test for didReceiveNotificationExtensionRequest contentHandler
Ignoring deprecated warnings for existing unit tests
The new unit test validates that the contentHandler is fired, and that the dispatch_semaphore is signaled within a second. We do not randomize the time sent for confirmed deliveries in unit tests.

The dispatch_semaphore now has a timeout of 30 seconds instead of "forever"