Checking OneSignal.isRegiusterUserFinished before displaying IAMs
If we display IAMs prior to receiving the response to on session, we could be displaying cached IAMs that have been paused, or edited. Additionally this is causing an issue where we display IAMs twice because of the self.messages = newMessages line in updateInAppMessagesFromOnSession. We overwrite the seenInSession variable of the old object with the new one. By waiting to display IAMs until onSession has returned we fix this issue. To avoid a case where cached IAMs aren't shown when onSession fails I am now calling updateInAppMessagesFromCache in the failure block of registerUserInternal.