Guarantee that messageView is a subview of OSInAppMessageVC
There is swiftUI crash where addConstraintsForMessage can be called without messageView being a subview of OSInAppMessageViewController's view.  This is likely happening because viewWillTransitionToSize is being called before setupInitialMessageUI. Neither the crash reporters nor OneSignal have been able to reproduce the crash, but it is safe to always add messageView as a subview if it is not already.