Tracking cold restarts
A cold restart is when an app is closed (swipe away) and relaunched within 30 seconds. In this scenario we currently won't call onSession. We want to track how often this occurs to know the impact of calling onSession in this scenario.

We don't want to consider backgrounding in this case so we are using the registerUserFinished boolean to remove cases where we have already registered a session. We also only want to track this once per app launch so the trackedColdRestart boolean prevents it from being called multiple times.