Putting try catch around fileHandle writeData
For iOS 13+ devices we will now use writeData:error: which returns a boolean and populates the NSError parameter with the failure reason if it  fails.

For < iOS 13 we will place a try catch around writeData: to avoid crashes.

In both cases we will log the failure reason. Note that this is logged on the NotificationServiceExtension process not the main app process.