add SwizzlingForwarder and test using it
Added one failing test for the case where forwardingTargetForSelector
is not accounted for in swizzling. The most common case for this is a
SwiftUI app with @UIApplicationDelegateAdaptor.

Just added one test and fixed one case to keep this PR small and
readable, will add the rest in a follow up commit.