Add double swizzle check
This fixes a bug where an infinite loop happens when we call the origin
implementation if delegates we swizzle if something attempts to set the
delegate back to the original class. This would only happen if the class
didn't have the target selector to begin with when we run this swizzle
logic, but then it is called a 2nd time with same target class. We have
a check in the setOneSignalDelegate and setOneSignalUNDelegate but if
the app developer or another library changes the delegate then puts it
back this scenario can happen.

Added a test for this scenario, ensuring it fails and then passing
after. This also fixes an issue in the last commit where this bug was
causing tests after UIApplicationDelegateSwizzingTest to infinitely
loop.