Fix testUNUserNotificationCenterDelegateAssignedBeforeOneSignal
This test broke in the last commit since the test was relying on
a 2nd call to injectSelector to unswizzle. This was intentionally done
as we don't want app or other SDKs swapping delegates back and forth to
cause OneSignal to no longer receive events.

This forced me to rewrite the OneSignalUNUserNotificationCenterHelper
used for tests to not reply on this behavior and in the process made
much simpler. It is declarative of the methods to swap back and
therefore much easier to understand.