add tests swizzling AppDelegate with inheritance
We have logic in our swizzling checking for parent classes however there
were no tests covering this.

testAppDelegateInheritsFromBaseWhereBothHaveSelectorsButSuperIsNotCalled
is a test that was added here that fails, uncovers a bug with our
swizzling!
The issue is injectToProperClass takes in a list of child classes
however it only swizzles the first one instead of all of them, which
we my swizzle one that isn't even in use. However this raises the
question why did we not keep it simple and just swizzle the passed in
delegate instance? That will be something to solve in the next commit!