rm unneeded parent and child class searches
The class that was passed in is the only thing that needs to be
swizzled, the checks to try to swizzle the parent class or child classes
was found to be unnecessary. In the last commit a number of tests
starting with testAppDelegateInheritsFromBase* were added to cover all
possible inheritance scenarios and these still pass after making this
change.
testAppDelegateInheritsFromBaseWhereBothHaveSelectorsButSuperIsNotCalled
was one test that was actually failing but now passes after this change.

Research was done to find the reason behind
getClassWithProtocolInHierarchy to see why it was added incase I
overlooked something. The root of the change was found in the following:
https://github.com/GameThrive/ios-push-plugin/commit/e656e99df690959ff82ec9715240dd50d360c090#diff-7ebd209501fe17a14bf53eaa7a246d9fe9f7b4b1d5139cdd2eb672fb3ae2fc29R900
Unfortunately the details are light, it notes that
"...a class that inherits from your AppDelegate class", however we have
added a unit test for this and things are working as expected. Since
this was added ~8 years ago so many things have change both in this code
base as well as iOS itself the original cause my no longer be an issue.

Since getClassWithProtocolInHierarchy was removed we no longer need
ClassGetSubclasses. This is because [delegate class] is the concrete
class.