clean up dead swizzling code
The last commit removed the usage of getClassWithProtocolInHierarchy and
ClassGetSubclasses in swizzling and this PR is cleaning up the code
itself. Was also able to remove checkIfInstanceOverridesSelector, since
it was only used in relation to these methods. injectToProperClass can
be removed, however will do so in a different comment since this will
touch a number of files.