replace injectToProperClass with injectSelector
Since the last commit no more rule is needed to for sub classes so
injectSelector was replace everywhere in the code base.

The parameters were reordered on injectSelector to make it more
readable. Starting with the original class and selector then followed by
the replacement class and selector is easier see what is targeting and
what it is changing to.