Ensure we only swizzle an AppDelegate class once
This was added to prevent an infinite loops when another library that
swizzles the same selectors is present in the app. This was only true
if the other library correctly forwards the event to the original
selector, however this is almost always the case as otherwise this other
SDK would create side effects if it missed the forwarding step.

Basically something can swizzle after OneSignal and swizzle it out of
first inline. However this is ok since that other swizzling library will
just think OneSignal's implementation is the original so therefore
OneSignal will still get the event. We just have trust the other SDK
swizzling is forwarding.