move location fetch to a background thread
running locationServicesEnabled on the main thread causes an Xcode warning in Xcode 14 about a potential UI holdup. However the location manager class needs to be created and run on a thread with a runloop or it's updates won't come through. To handle this we will do locationServicesEnabled on a background thread and location fetch on main thread