add onJwtExpired method
This method takes a closure from the developer. The closure will be called when the Jwt token expires. The closure takes the externalId that is associated with the expired token as well as a completion block. The completion block should be called the app developer with the new jwt token for that user.