Adding 2 new github actions that allow us to measure issue response time
One action is triggered when an issue is closed or it is commented on by a OneSignal employee. This action sets the response time in business days for that issue inside the issue html as a comment. This is based on https://github.com/probot/metadata

The other action is manually dispatched and calculates the average issue reponse time for the repo for issues created from some start date to some end date.

It also says how many issues have not been responded to.