Removing launchURLsInApp and webkit code from Notifications
We are removing the launch push notification urls in a webview feature. This feature was not consistently effective, and added a webkit dependency to notifications that has gotten apps rejected from the app store. If we later want to add this feature back we should consider it as an optional module.