Always set test type to 1 for simulators
Cherry-pick of #1244

Our provision detection in OneSignalMobileProvision is not successfully able to find the embedded mobile provision for simulators. This might only affect Xcode managed provisioning profiles.

This change is a workaround to that issue to always set testType to 1 since simulators need to receive push in the sandbox environment.