Make web views inspectable in debug log mode or greater
We build using iOS 14.1 which does not support webview.inspectable = true

We will use reflection to only call the method if the device supports it. Additionally since the method has a boolean argument we must use NSInvocation instead of performSelector

I also split the @available check with the debug check since they cannot be in the same statement.