Add user state observer
Problem:
App developers want to know the onesignal ID for a user, but it can be null if we have not received the response from the server after a fresh install or after logging in.

Solution:
Offer a user state observer that will fire when the response is received and fire with the onesignal ID and external ID.

Implementation Details:
On Identity Model hydration from the server (when onesignal_id is received), we will snapshot the onesignal_id and external_id as the conclusive IDs for the current user. On future hydrations after logging in, we will compare to the previous snapshot to know if the observer should be fired.

We must not only compare onesignal_id but also external_id for the case of logging in Anonymous -> New Identified. In this case, the onesignal_id will not change but the external_id will be attached to the user on the server.